+++ domain/repositories/audit_log_repository.go
@@ -6,8 +6,8 @@ import (
 
 type AuditLogRepository interface {
 	Create(auditLog *entity.AuditLog) error
-	GetByID(id uint) (*entity.AuditLog, error)
-	GetByOperator(operatorID uint, page, pageSize int) ([]entity.AuditLog, int64, error)
-	GetByTarget(targetType string, targetID uint, page, pageSize int) ([]entity.AuditLog, int64, error)
+	GetByID(id uint64) (*entity.AuditLog, error)
+	GetByOperator(operatorID uint64, page, pageSize int) ([]entity.AuditLog, int64, error)
+	GetByTarget(targetType string, targetID uint64, page, pageSize int) ([]entity.AuditLog, int64, error)
 	GetByTimeRange(startTime, endTime string, page, pageSize int) ([]entity.AuditLog, int64, error)
 }
