+++ domain/repositories/org_unit_repository.go
@@ -7,9 +7,9 @@ import (
 type OrgUnitRepository interface {
 	Create(orgUnit *entity.OrgUnit) error
 	Update(orgUnit *entity.OrgUnit) error
-	Delete(id uint) error
-	GetByID(id uint) (*entity.OrgUnit, error)
+	Delete(id uint64) error
+	GetByID(id uint64) (*entity.OrgUnit, error)
 	GetByPath(path string) (*entity.OrgUnit, error)
 	GetAll() ([]entity.OrgUnit, error)
-	GetChildren(parentID uint) ([]entity.OrgUnit, error)
+	GetChildren(parentID uint64) ([]entity.OrgUnit, error)
 }
