+++ domain/services/org_unit_service.go
@@ -5,10 +5,10 @@ import (
 )
 
 type OrgUnitService interface {
-	CreateOrgUnit(orgUnit *entity.OrgUnit, operatorID uint) error
-	UpdateOrgUnit(orgUnit *entity.OrgUnit, operatorID uint) error
-	DeleteOrgUnit(id uint, operatorID uint) error
-	GetOrgUnitByID(id uint) (*entity.OrgUnit, error)
+	CreateOrgUnit(orgUnit *entity.OrgUnit, operatorID uint64) error
+	UpdateOrgUnit(orgUnit *entity.OrgUnit, operatorID uint64) error
+	DeleteOrgUnit(id uint64, operatorID uint64) error
+	GetOrgUnitByID(id uint64) (*entity.OrgUnit, error)
 	GetAllOrgUnits() ([]entity.OrgUnit, error)
 	GetOrgTree() ([]entity.OrgUnit, error)
 }
